# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Project & Task API",
        "description": "API RESTful para la gestión de proyectos y tareas con Laravel 12",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor local de desarrollo"
        }
    ],
    "paths": {
        "/api/projects": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "Listar proyectos con filtros dinámicos",
                "operationId": "fbc21feeb3df8a2cb519efb6480ae7d9",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filtrar por estado (active, inactive)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Buscar por nombre",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "from",
                        "in": "query",
                        "description": "Fecha inicio rango (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "description": "Fecha fin rango (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de proyectos paginada"
                    }
                }
            },
            "post": {
                "tags": [
                    "Projects"
                ],
                "summary": "Crear un nuevo proyecto",
                "operationId": "a9d5bd5914cf75eaebafcf8b9e320dcf",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "status"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Proyecto API"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Proyecto de ejemplo"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "active"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Proyecto creado correctamente"
                    }
                }
            }
        },
        "/api/projects/{id}": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "Mostrar detalle de un proyecto",
                "operationId": "865085b82af27d647cf1f2a772aa7927",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del proyecto",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalle del proyecto"
                    }
                }
            },
            "put": {
                "tags": [
                    "Projects"
                ],
                "summary": "Actualizar un proyecto existente",
                "operationId": "fe600b7ffce78456bce67c51b9eca51d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del proyecto",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "status"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Proyecto Actualizado"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Descripción actualizada"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "inactive"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Proyecto actualizado correctamente"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Projects"
                ],
                "summary": "Eliminar un proyecto",
                "operationId": "d81ac2b0b51fa7143c69f1a5f4c27eb0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del proyecto",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Proyecto eliminado correctamente"
                    }
                }
            }
        },
        "/api/tasks": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Listar tareas con filtros dinámicos",
                "operationId": "be68905d866e856db6c738172b8d929a",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filtrar por estado (pending, in_progress, done)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "priority",
                        "in": "query",
                        "description": "Filtrar por prioridad (low, medium, high)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "due_date",
                        "in": "query",
                        "description": "Filtrar por fecha de entrega",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "project_id",
                        "in": "query",
                        "description": "Filtrar por ID de proyecto",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de tareas paginada"
                    }
                }
            },
            "post": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Crear nueva tarea",
                "operationId": "ebbaabc58e3db61374a57c53544959b2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "project_id",
                                    "title",
                                    "status",
                                    "priority",
                                    "due_date"
                                ],
                                "properties": {
                                    "project_id": {
                                        "type": "string",
                                        "example": "uuid-proyecto"
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "Configurar entorno"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Instalar dependencias"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "pending"
                                    },
                                    "priority": {
                                        "type": "string",
                                        "example": "high"
                                    },
                                    "due_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-08-25"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tarea creada correctamente"
                    }
                }
            }
        },
        "/api/tasks/{id}": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Mostrar detalle de una tarea",
                "operationId": "32c40e2d1129b80c1dcabd538df63a33",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la tarea",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalle de la tarea"
                    }
                }
            },
            "put": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Actualizar una tarea",
                "operationId": "c36c959b634c99e1e02ad2f3f2c36a2c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la tarea",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "status",
                                    "priority",
                                    "due_date"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Configurar entorno actualizado"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Actualizar dependencias"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "in_progress"
                                    },
                                    "priority": {
                                        "type": "string",
                                        "example": "medium"
                                    },
                                    "due_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-08-30"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tarea actualizada correctamente"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Eliminar una tarea",
                "operationId": "d81cc7975ea62176b4ef8f74ed95168d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la tarea",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tarea eliminada correctamente"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Projects",
            "description": "Projects"
        },
        {
            "name": "Tasks",
            "description": "Tasks"
        }
    ]
}